+++ frontend/src/components/Player/Player.tsx
@@ -1,18 +1,14 @@
 import { useSocketStore } from '@/stores';
-import { css, keyframes } from '@emotion/react';
+import { css } from '@emotion/react';
 import { useRef, useState } from 'react';
 import { divideSize, multiplySize } from '@/utils';
-import { Variables } from '@/styles';
 import ReactPlayer from 'react-player';
-import PlayIcon from '@/assets/icons/play-fill.svg?react';
-import PauseIcon from '@/assets/icons/pause-line.svg?react';
 import { useFraction, useToast } from '@/hooks';
-import { InterestYoutubeResponse, YoutubeRequestType } from '@/types';
+import { InterestYoutubeResponse, PlayerState, YoutubeRequestType } from '@/types';
 import { YOUTUBE_ERROR_MESSAGES } from '@/constants';
 import SharerDraggingIndicator from './SharerDraggingIndicator';
 import ControllerSection from './ControllerSection';
-
-type StateChange = 'pause' | 'play';
+import StateChangeIndicator from './StateChangeIndicator';
 
 interface PlayerProps {
   url: string;
@@ -25,7 +21,7 @@ const PLAYER_WIDTH_DEFAULT = '40rem';
 const PLAYER_HEIGHT_DEFAULT = divideSize(PLAYER_WIDTH_DEFAULT, 16 / 9);
 const PLAYER_HEIGHT_DOUBLE = multiplySize(PLAYER_HEIGHT_DEFAULT, 2);
 
-const statusEventNameMap: Record<StateChange, string> = {
+const statusEventNameMap: Record<PlayerState, string> = {
   play: 'interest:youtube:play',
   pause: 'interest:youtube:stop'
 };
@@ -45,38 +41,6 @@ const mediaSectionStyle = css({
   width: '100%',
   height: '80%'
 });
-
-const growAndDiplay = (baseTransform: string) =>
-  keyframes({
-    '0%': {
-      opacity: '0',
-      transform: `${baseTransform} scale(0)`
-    },
-    '48%': {
-      opacity: '1',
-      transform: `${baseTransform} scale(1)`
-    },
-    '60%': {
-      opacity: '0.3',
-      transform: `${baseTransform} scale(1.4)`
-    },
-    '100%': {
-      opacity: '0',
-      transform: `${baseTransform} scale(2)`
-    }
-  });
-
-const stateChangeIndicatorStyle = (stateChanged: boolean) =>
-  css({
-    position: 'absolute',
-    top: '50%',
-    left: '50%',
-    opacity: '0',
-    animation: stateChanged ? `${growAndDiplay('translate(-50%, -50%)')} 0.5s ease-out` : '',
-    fill: Variables.colors.surface_white,
-    width: '3rem',
-    zIndex: '999'
-  });
 
 const Player = ({ url, isSharer, isShorts }: PlayerProps) => {
   const [isPlaying, setIsPlaying] = useState(true);
@@ -103,17 +67,17 @@ const Player = ({ url, isSharer, isShorts }: PlayerProps) => {
     DRAGGING: syncWithSharerDragStart
   };
 
-  function playVideo() {
+  function playVideoAsSharer() {
     if (!isSharer) return;
     prevIsPlayingRef.current = true;
 
     setIsPlaying(true);
     player?.getInternalPlayer().playVideo();
 
-    sendStateChange('play');
+    sendPlayerState('play');
   }
 
-  function pauseVideo() {
+  function pauseVideoAsSharer() {
     if (!isSharer) return;
     prevIsPlayingRef.current = false;
 
@@ -121,12 +85,12 @@ const Player = ({ url, isSharer, isShorts }: PlayerProps) => {
 
     player?.getInternalPlayer().pauseVideo();
 
-    sendStateChange('pause');
+    sendPlayerState('pause');
   }
 
   function toggleVideo() {
-    if (isPlaying) pauseVideo();
-    else playVideo();
+    if (isPlaying) pauseVideoAsSharer();
+    else playVideoAsSharer();
   }
 
   function syncWithSharerPlayOrPause({
@@ -201,7 +165,7 @@ const Player = ({ url, isSharer, isShorts }: PlayerProps) => {
     });
   }
 
-  function sendStateChange(stateChange: StateChange) {
+  function sendPlayerState(stateChange: PlayerState) {
     const eventName = statusEventNameMap[stateChange];
     const body: Record<string, any> = { videoCurrentTime: player?.getCurrentTime(), playStatus: stateChange };
 
@@ -253,13 +217,7 @@ const Player = ({ url, isSharer, isShorts }: PlayerProps) => {
             }}
           ></div>
         )}
-
-        {isPlaying ? (
-          <PlayIcon css={stateChangeIndicatorStyle(prevIsPlayingRef.current === isPlaying)} />
-        ) : (
-          <PauseIcon css={stateChangeIndicatorStyle(prevIsPlayingRef.current === isPlaying)} />
-        )}
-
+        <StateChangeIndicator isPlaying={isPlaying} prevIsPlayingRef={prevIsPlayingRef} />
         <ReactPlayer
           style={{
             pointerEvents: 'none',
@@ -281,7 +239,7 @@ const Player = ({ url, isSharer, isShorts }: PlayerProps) => {
             if (isDraggingSliderRef.current) return;
             syncFractionWithProgress({ played: 1 });
             hasEndedRef.current = true;
-            pauseVideo();
+            pauseVideoAsSharer();
           }}
           config={{ youtube: { playerVars: { autoplay: 1 } } }}
         />
@@ -291,8 +249,8 @@ const Player = ({ url, isSharer, isShorts }: PlayerProps) => {
           player={player}
           isSharer={isSharer}
           isPlaying={isPlaying}
-          playVideo={playVideo}
-          pauseVideo={pauseVideo}
+          playVideo={playVideoAsSharer}
+          pauseVideo={pauseVideoAsSharer}
           setFraction={setFraction}
           setVolume={setVolume}
           volume={volume}
