+++ frontend/src/styles/index.ts
@@ -1,4 +1,4 @@
-export { rotate, growAndShrink, hoverGrowJumpAnimation, fadeIn, fadeOut, scaleIn } from './animation';
+export { rotate, growAndShrink, hoverGrowJumpAnimation, fadeIn, fadeOut, scaleIn, growAndDiplay } from './animation';
 export { flexStyle } from './flex';
 export { Variables } from './Variables';
 export { keywordStyleMap } from './keywordStyleMap';
