+++ frontend/src/types/index.ts
@@ -12,3 +12,4 @@ export type {
 } from './response';
 export type { CommonResult } from './commonResult';
 export type { Content } from './content';
+export type { PlayerState } from './videoPlayer';
