+++ frontend/src/types/videoPlayer.ts
@@ -1,0 +1,1 @@
+export type PlayerState = 'pause' | 'play';
